Changed types of data members to size_t.

+++ Ship.cpp
@@ -2,13 +2,13 @@
 
 Ship::Ship() = default;
 
-Ship::Ship(int id, const std::string& name, unsigned int speed, unsigned int maxCrew, unsigned int capacity)
+Ship::Ship(int id, const std::string& name, size_t speed, size_t maxCrew, size_t capacity)
     : id_(id), name_(name), speed_(speed), maxCrew_(maxCrew), capacity_(capacity)
 {
 
 }
 
-Ship::Ship(int id, unsigned int speed, unsigned int maxCrew)
+Ship::Ship(int id, size_t speed, size_t maxCrew)
     : Ship(id, "", speed, maxCrew, 1)
 {
